+++ Js2IL.Tests/Promise/JavaScript/Promise_WithResolvers_Idempotent.js
@@ -1,0 +1,8 @@
+const r = Promise.withResolvers();
+
+r.promise
+    .then((value) => console.log("then:" + value))
+    .catch((reason) => console.log("catch:" + reason));
+
+r.resolve("first");
+r.reject("second");