+++ Js2IL.Tests/Promise/JavaScript/Promise_WithResolvers_Resolve.js
@@ -1,0 +1,7 @@
+const r = Promise.withResolvers();
+
+r.promise.then((value) => {
+    console.log(value);
+});
+
+r.resolve("resolved");