+++ Js2IL/IR/HIR/HIRAwaitExpression.cs
@@ -1,0 +1,11 @@
+namespace Js2IL.HIR;
+
+public sealed class HIRAwaitExpression : HIRExpression
+{
+    public HIRAwaitExpression(HIRExpression argument)
+    {
+        Argument = argument;
+    }
+
+    public HIRExpression Argument { get; }
+}